Convert to Objective-C (main.c -> main.m)
This will enabled me to use NSAttributedString, which seems to be the
easiest way to deal with RTF data (which is what the "Secure Notes"
actually are).

It will also enabled me to use some NS* datatypes along with toll-free
bridging to make dealing with some of these CF* APIs a little less
painful.

Signed-off-by: Greg Hurrell <greg@hurrell.net>